requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-telegram-bot
aiogram
streamlit
requests
pandas
